pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anki-vocab"
version = "0.1.0"
description = "Generate Anki cards from BCS vocabulary using OpenAI and Pexels"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = "^3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "structlog>=23.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.0",
    "vcrpy>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "vcrpy>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
anki-vocab = "anki_vocab.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["anki_vocab"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
markers = [
    "live_llm: marks tests as live LLM tests (deselect with '-m \"not live_llm\"')",
] 